+++ 1.21.5/CHANGELOG.md
@@ -4,9 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [v21.5.5-1.21.5] - 2025-04-05
+## [v21.5.5-1.21.5] - 2025-04-10
 ### Added
 - Add `ClientCommonHelper`
+- Add `GuiHeightHelper::getMaxHeight` & `GuiHeightHelper::getMinHeight`
+### Fixed
+- Fix `ClassCastException` for `PlayLevelSoundEvents` on Fabric
 
 ## [v21.5.4-1.21.5] - 2025-04-03
 ### Added